Refactor code strucutre.
