+++ src/main/scala/com/treadstone90/mesos/http/MasterClient.scala
@@ -1,4 +1,4 @@
-package http
+package com.treadstone90.mesos.http
 
 /**
   * Created by karthik on 7/16/17.
