+++ src/main/scala/com/treadstone90/mesos/http/MesosMasterHttpClientFactory.scala
@@ -1,7 +1,9 @@
-package http
+package com.treadstone90.mesos.http
 
 import com.google.common.net.HostAndPort
-import mesos.{Driver, MesosEventHandler}
+import com.treadstone90.mesos.MesosEventHandler
+import com.treadstone90.mesos.scheduler.Driver
+import example.MesosEventHandler
 import org.apache.mesos.v1.mesos.{FrameworkInfo, MasterInfo}
 
 /**
