--- src/main/scala/http/DriverNotStartedException.scala
@@ -1,6 +1,0 @@
-package http
-
-/**
-  * Created by karthik on 7/20/17.
-  */
-case class DriverNotStartedException(msg: String) extends Exception
